검색 창 색상 지정, 스크롤 컴포넌트 css 수정 홈 상단 폴더, 파일 추가 및 수정
